Check test randomness outcome is deterministic with --random-order-seed set
