+++ test/Unit/Runners/PHPUnit/BaseRunnerTest.php
@@ -4,20 +4,12 @@ declare(strict_types=1);
 
 namespace ParaTest\Tests\Unit\Runners\PHPUnit;
 
-use ParaTest\Runners\PHPUnit\Options;
 use ParaTest\Tests\TestBase;
-use Symfony\Component\Process\Process;
 
-use function array_reverse;
 use function assert;
 use function count;
-use function file_get_contents;
 use function glob;
-use function posix_mkfifo;
-use function preg_match_all;
-use function preg_quote;
 use function simplexml_load_file;
-use function sprintf;
 
 /**
  * @internal
@@ -168,135 +160,5 @@ final class BaseRunnerTest extends TestBase
         static::assertIsArray($attribues['@attributes']);
         static::assertArrayHasKey('name', $attribues['@attributes']);
         static::assertSame('', $attribues['@attributes']['name']);
-    }
-
-    public function testTeamcityLog(): void
-    {
-        $outputPath = TMP_DIR . DS . 'test-output.teamcity';
-
-        $this->bareOptions = [
-            '--configuration' => $this->fixture('phpunit-passing.xml'),
-            '--log-teamcity' => $outputPath,
-        ];
-
-        $this->runRunner();
-
-        static::assertFileExists($outputPath);
-        $content = file_get_contents($outputPath);
-        static::assertNotFalse($content);
-
-        self::assertSame(66, preg_match_all('/^##teamcity/m', $content));
-    }
-
-    /**
-     * @requires OSFAMILY Linux
-     */
-    public function testTeamcityLogHandlesFifoFiles(): void
-    {
-        $outputPath = TMP_DIR . DS . 'test-output.teamcity';
-
-        posix_mkfifo($outputPath, 0600);
-        $this->bareOptions = [
-            '--configuration' => $this->fixture('phpunit-passing.xml'),
-            '--log-teamcity' => $outputPath,
-        ];
-
-        $fifoReader = new Process(['cat', $outputPath]);
-        $fifoReader->start();
-
-        $this->runRunner();
-
-        self::assertSame(0, $fifoReader->wait());
-        self::assertSame(66, preg_match_all('/^##teamcity/m', $fifoReader->getOutput()));
-    }
-
-    public function testRunnerSort(): void
-    {
-        $this->bareOptions = [
-            '--order-by' => Options::ORDER_RANDOM,
-            '--random-order-seed' => 123,
-            '--configuration' => $this->fixture('phpunit-passing.xml'),
-        ];
-
-        $runnerResult = $this->runRunner();
-
-        static::assertStringContainsString('Random order seed 123', $runnerResult->getOutput());
-    }
-
-    public function testRunnerSortNoSeedProvided(): void
-    {
-        $this->bareOptions = [
-            '--order-by' => Options::ORDER_RANDOM,
-            '--configuration' => $this->fixture('phpunit-passing.xml'),
-        ];
-
-        $runnerResult = $this->runRunner();
-
-        static::assertStringContainsString('Random order seed ', $runnerResult->getOutput());
-    }
-
-    public function testRunnerSortTestEqualBySeed(): void
-    {
-        $this->bareOptions = [
-            '--configuration' => $this->fixture('phpunit-passing.xml'),
-            '--order-by' => Options::ORDER_RANDOM,
-            '--random-order-seed' => 123,
-            '--verbose' => 1,
-        ];
-
-        $runnerResultFirst  = $this->runRunner();
-        $runnerResultSecond = $this->runRunner();
-
-        $firstOutput  = $this->prepareOutputForTestOrderCheck($runnerResultFirst->getOutput());
-        $secondOutput = $this->prepareOutputForTestOrderCheck($runnerResultSecond->getOutput());
-        static::assertSame($firstOutput, $secondOutput);
-
-        $this->bareOptions['--random-order-seed'] = 321;
-
-        $runnerResultThird = $this->runRunner();
-
-        $thirdOutput = $this->prepareOutputForTestOrderCheck($runnerResultThird->getOutput());
-
-        static::assertNotSame($thirdOutput, $firstOutput);
-    }
-
-    public function testRunnerReversed(): void
-    {
-        $this->bareOptions = [
-            '--configuration' => $this->fixture('phpunit-passing.xml'),
-            '--verbose' => 1,
-        ];
-
-        $runnerResult = $this->runRunner();
-        $defaultOrder = $this->prepareOutputForTestOrderCheck($runnerResult->getOutput());
-
-        $this->bareOptions['--order-by'] = Options::ORDER_REVERSE;
-
-        $runnerResult = $this->runRunner();
-        $reverseOrder = $this->prepareOutputForTestOrderCheck($runnerResult->getOutput());
-
-        $reverseOrderReversed = array_reverse($reverseOrder);
-
-        static::assertSame($defaultOrder, $reverseOrderReversed);
-    }
-
-    /**
-     * @return string[]
-     */
-    private function prepareOutputForTestOrderCheck(string $output): array
-    {
-        $matchesCount = preg_match_all(
-            sprintf(
-                '/%s%s(?<filename>\S+\.php)/',
-                preg_quote(FIXTURES, '/'),
-                preg_quote(DS, '/')
-            ),
-            $output,
-            $matches
-        );
-
-        self::assertGreaterThan(0, $matchesCount);
-
-        return $matches['filename'];
     }
 }
