+++ test/fixtures/phpunit-deterministic-random.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<phpunit>
+    <testsuites>
+        <testsuite name="ParaTest Fixtures">
+            <directory>./deterministic_random/</directory>
+        </testsuite>
+    </testsuites>
+</phpunit>
